+++ apps/ui-dev/component/navigations/treeview/MainView.jsx
@@ -2,7 +2,13 @@ import ChevronRightFilled from "@elf-framework/icon/ChevronRightFilled";
 import DescriptionFilled from "@elf-framework/icon/DescriptionFilled";
 import FolderFilled from "@elf-framework/icon/FolderFilled";
 import { useRender, useState } from "@elf-framework/sapa";
-import { Button, TreeView, TreeViewProvider, VBox } from "@elf-framework/ui";
+import {
+  Button,
+  FixedTooltip,
+  TreeView,
+  TreeViewProvider,
+  VBox,
+} from "@elf-framework/ui";
 
 import treeData from "./data/base.json";